requirements.txt
asgiref==3.7.2
boto3==1.28.24
botocore==1.31.24
dj-database-url==2.0.0
Django==4.2.4
django-crispy-forms==1.14.0
django-jazzmin==2.6.0
django-shortcuts==1.6
django-storages==1.13.2
gunicorn==21.2.0
jmespath==1.0.1
packaging==23.1
psycopg2-binary==2.9.7
python-dateutil==2.8.2
python-dotenv==1.0.0
s3transfer==0.6.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
